+++ SOC_Particle/src/Cloud.h
@@ -18,7 +18,7 @@ struct Publish
   float Amp_hrs_remaining_soc;
 };
 
-void assign_publist(Publish* pubList, const unsigned long now, const String unit, const String hm_string,
+void assign_publist(Publish* pubList, const unsigned long long now, const String unit, const String hm_string,
   Sensors* Sen, const int num_timeouts, BatteryMonitor* Mon);
 
 #endif
